Fix
Configured openvpn systemd service to start as user tunnel. Otherwise openvpn
would initially start as root, connect and then drop to and continue to connect
as user tunnel. This would not work, because the VPN_FIREWALL allows only user
tunnel to connect to external destinations. Added (and out commented by
default) sudoers exception to allow user tunnel to run openvpn as root so it
can delete/add the tun device.

https://phabricator.whonix.org/T460